requirements.txt
python-dateutil
requests
discord_webhook
python-dotenv